\name{chr20}
\alias{chr20}
\alias{mhc}
\docType{data}
\title{DNA datasets}
\description{
Example DNA sequences.
}
\usage{data(chr20)}
\details{
The datasets are the chromosome 20 sequence of the human genome, and the mhc dataset available from the Human Genome Project website, binary-coded by base pair content and curtailed to a power of two. 
}
\source{
\url{http://www.sanger.ac.uk/HGP/} 
}
\keyword{datasets}
