\name{ebayesthresh.wavelet.wd}
\alias{ebayesthresh.wavelet.wd}
\alias{threshold.wd}
\alias{negloglik.laplace}
\alias{wandafromx}
\alias{madmad}
\title{Modified EbayesThresh wavelet thresholding function}
\description{
Modified EbayesThresh functions. 
}
%\usage{
%ebayesthresh.wavelet.wd(x.wd, vscale = "independent", smooth.levels = Inf, 
%    prior = "laplace", a = 0.5, bayesfac = FALSE, threshrule = "median")
%}
%\arguments{
%  \item{x.wd}{see original help file.}
%  \item{vscale}{see original help file.}
%  \item{smooth.levels}{see original help file.}
%  \item{prior}{see original help file.}
%  \item{a}{see original help file.}
%  \item{bayesfac}{see original help file.}
%  \item{threshrule}{see original help file.}
%  }
\details{
  For help on these function, see the original help file supplied with the WaveThresh package.  There is a modification to try and avoid zero noise standard deviation estimation.
}
\keyword{regression}% at least one, from doc/KEYWORDS
